rework spio tests (not working)
