+++ src/axom/sidre/tests/spio/CMakeLists.txt
@@ -11,12 +11,12 @@
 #------------------------------------------------------------------------------
 
 set(gtest_spio_serial_tests
-   spio_serial.cpp
+   spio_serial.hpp
    )
 
 set(gtest_spio_parallel_tests
-   spio_basic.cpp
-   spio_parallel.cpp
+   spio_basic.hpp
+   spio_parallel.hpp
    )
 
 set(spio_test_depends
@@ -28,108 +28,90 @@ set(spio_test_depends
 blt_list_append(TO spio_test_depends ELEMENTS hdf5 IF HDF5_FOUND)
 
 if(SCR_FOUND)
-     list(APPEND spio_test_depends scr)
-     list(APPEND gtest_spio_parallel_tests spio_scr.cpp)
+    list(APPEND spio_test_depends scr)
+    list(APPEND gtest_spio_parallel_tests spio_scr.hpp)
 endif()
 
 #------------------------------------------------------------------------------
 # Add gtest C++ tests
 #------------------------------------------------------------------------------
-if (ENABLE_MPI)
-
-   foreach(test ${gtest_spio_serial_tests})
-       get_filename_component( test_name ${test} NAME_WE )
-       blt_add_executable( NAME ${test_name}_test
-                           SOURCES ${test}
-                           OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                           DEPENDS_ON ${spio_test_depends} gtest
-                           FOLDER axom/sidre/tests
-                           )
-
-       blt_add_test( NAME ${test_name} 
-                     COMMAND ${test_name}_test
-                     NUM_MPI_TASKS 1
+blt_add_executable( NAME       spio_tests
+                    SOURCES    spio_main.cpp ${gtest_spio_parallel_tests} ${gtest_spio_serial_tests}
+                    OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
+                    DEPENDS_ON ${spio_test_depends} gtest
+                    FOLDER     axom/sidre/tests
                     )
-   endforeach()
 
+foreach(test ${gtest_spio_serial_tests})
+    get_filename_component( test_name ${test} NAME_WE )
+    blt_add_test( NAME          ${test_name} 
+                  COMMAND       spio_tests --gtest_filter=${test_name}*
+                  NUM_MPI_TASKS 1
+                 )
+endforeach()
 
-   foreach(test ${gtest_spio_parallel_tests})
-       get_filename_component( test_name ${test} NAME_WE )
-       blt_add_executable( NAME ${test_name}_test
-                           SOURCES ${test}
-                           OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                           DEPENDS_ON ${spio_test_depends} gtest
-                           FOLDER axom/sidre/tests
-                           )
-
-       blt_add_test( NAME ${test_name} 
-                     COMMAND ${test_name}_test
-                     NUM_MPI_TASKS 4
-                     )
-   endforeach()
-
-endif()
+foreach(test ${gtest_spio_parallel_tests})
+    get_filename_component( test_name ${test} NAME_WE )
+    blt_add_test( NAME          ${test_name} 
+                  COMMAND       spio_tests --gtest_filter=${test_name}*
+                  NUM_MPI_TASKS 4
+                  )
+endforeach()
 
 #------------------------------------------------------------------------------
 # Add Fortran tests
 #------------------------------------------------------------------------------
 if(ENABLE_FORTRAN)
-   if (ENABLE_MPI)
-      blt_add_executable(NAME spio_basicWriteRead_F_test
-                         SOURCES F_spio_basicWriteRead.F
-                         OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                         DEPENDS_ON ${spio_test_depends}
-                         FOLDER axom/sidre/tests
-                         )
-      blt_add_test(NAME F_spio_basicWriteRead
-                   COMMAND spio_basicWriteRead_F_test
-                   NUM_MPI_TASKS 1
-                   )
+    blt_add_executable(NAME spio_basicWriteRead_F_test
+                       SOURCES F_spio_basicWriteRead.F
+                       OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
+                       DEPENDS_ON ${spio_test_depends}
+                       FOLDER axom/sidre/tests
+                       )
+    blt_add_test(NAME F_spio_basicWriteRead
+                 COMMAND spio_basicWriteRead_F_test
+                 NUM_MPI_TASKS 1
+                 )
 
-      blt_add_executable(NAME spio_parallelWriteRead_F_test
-                         SOURCES F_spio_parallelWriteRead.F
-                         OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                         DEPENDS_ON ${spio_test_depends}
-                         FOLDER axom/sidre/tests
-                         )
-      blt_add_test(NAME F_spio_parallelWriteRead
-                   COMMAND spio_parallelWriteRead_F_test
-                   NUM_MPI_TASKS 4
-                   )
+    set(spio_fortran_tests
+        F_spio_blueprintIndex.F  
+        F_spio_parallelWriteRead.F
+        F_spio_irregularWriteRead.F
+        F_spio_externalWriteRead.F
+        )
 
-      blt_add_executable(NAME spio_irregularWriteRead_F_test
-                         SOURCES F_spio_irregularWriteRead.F
-                         OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                         DEPENDS_ON ${spio_test_depends}
-                         FOLDER axom/sidre/tests
-                         )
-      blt_add_test(NAME F_spio_irregularWriteRead
-                   COMMAND spio_irregularWriteRead_F_test
-                   NUM_MPI_TASKS 4
-                   )
+    foreach(test ${spio_fortran_tests})
+        get_filename_component( test_name ${test} NAME_WE )
+        string(SUBSTRING "${test_name}" 2 -1 test_name) # remove F_
+        blt_add_executable( NAME       ${test_name}_F_test
+                            SOURCES    ${test}
+                            OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
+                            DEPENDS_ON ${spio_test_depends}
+                            FOLDER     axom/sidre/tests
+                            )
+        blt_add_test( NAME    ${test_name}
+                      COMMAND ${test_name}_F_test
+                      NUM_MPI_TASKS 4
+                      )
+    endforeach()
 
-      blt_add_executable(NAME spio_externalWriteRead_F_test
-                         SOURCES F_spio_externalWriteRead.F
-                         OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                         DEPENDS_ON ${spio_test_depends}
-                         FOLDER axom/sidre/tests
-                         )
-      blt_add_test(NAME F_spio_externalWriteRead
-                   COMMAND spio_externalWriteRead_F_test
-                   NUM_MPI_TASKS 4
-                   )
+    set(spio_fruit_tests
+        F_spio_preserveWriteRead.F
+        )
 
-      blt_add_executable(NAME spio_preserveWriteRead_F_test
-                         SOURCES F_spio_preserveWriteRead.F
-                         OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
-                         DEPENDS_ON ${spio_test_depends} fruit
-                         FOLDER axom/sidre/tests
-                         )
-      blt_add_test(NAME F_spio_preserveWriteRead
-                   COMMAND spio_preserveWriteRead_F_test
-                   NUM_MPI_TASKS 4
-                   )
+    foreach(test ${spio_fruit_tests})
+        get_filename_component( test_name ${test} NAME_WE )
+        string(SUBSTRING "${test_name}" 2 -1 test_name) # remove F_
+        blt_add_executable( NAME       ${test_name}_F_test
+                            SOURCES    ${test}
+                            OUTPUT_DIR ${TEST_OUTPUT_DIRECTORY}
+                            DEPENDS_ON ${spio_test_depends} fruit
+                            FOLDER     axom/sidre/tests
+                            )
+        blt_add_test( NAME    ${test_name}
+                      COMMAND ${test_name}_F_test
+                      NUM_MPI_TASKS 4
+                      )
+    endforeach()
 
-   endif()
 endif()
-
