+++ src/axom/sidre/tests/spio/spio_basic.hpp
@@ -125,22 +125,3 @@ TEST(spio_basic, baton)
     checkBaton(nFiles, num_ranks, my_rank);
   }
 }
-
-//----------------------------------------------------------------------
-#include "axom/slic/core/UnitTestLogger.hpp"
-using axom::slic::UnitTestLogger;
-
-int main(int argc, char* argv[])
-{
-  int result = 0;
-
-  ::testing::InitGoogleTest(&argc, argv);
-
-  UnitTestLogger logger;  // create & initialize test logger,
-
-  MPI_Init(&argc, &argv);
-  result = RUN_ALL_TESTS();
-  MPI_Finalize();
-
-  return result;
-}
